requirements.txt
flask>=2.0.1
nltk>=3.6.1
python-twitter>=3.5
python-dotenv>=0.19.0